+++ micro/src/org/cougaar/microedition/tini/TiniMotorController.java
@@ -155,7 +155,7 @@ public class TiniMotorController extends MotorControllerResource {
     }
   }
 
-  public void getValues(double [] values)
+  public void getValues(long [] values)
   {
   }
